update confirmation ui
