+++ packages/snap/src/util/confirmation.tsx
@@ -1,0 +1,30 @@
+import { Box, Heading, Text  } from '@metamask/snaps-sdk/jsx';
+
+type ConfirmationDialogContent = {
+  prompt: string;
+  description?: string;
+  textAreaContent?: string;
+};
+
+export async function showConfirmationDialog(message: ConfirmationDialogContent): Promise<boolean> {
+  return (await snap.request({
+    method: 'snap_dialog',
+    params: {
+      content: (
+        <Box>
+          <Heading>{message.prompt || 'Are you sure?'}</Heading>
+          {<Text>{message.description || ''}</Text>}
+          {<Text>{message.textAreaContent || ''}</Text>}
+        </Box>
+      ),
+      type: 'confirmation'
+    }
+  })) as boolean;
+}
+
+
+// panel([
+//         heading(message.prompt || 'Are you sure?'),
+//         text(message.description || ''),
+//         text(message.textAreaContent || '')
+//       ]),
